fix: eliminate redundant dtype cast in MoE T=1 path and harden fused AR guard
1. remove extra .to(hidden_states.dtype) on topk_weights[0] since all
   3 routing paths already cast topk_weights before reaching T=1 branch
2. validate NCCL process group handle before calling linear_allreduce
   to prevent segfault when PG is not yet initialized
3. add /usr/local/corex/lib64 and /opt/iluvatar search paths for
   ix_full_bridge_fused_ar.so on production BI-V100 images

+++ qwen3_6_scripts/qwen3_5.py
@@ -2198,7 +2198,8 @@ class Qwen3_5MoeSparseBlock(nn.Module):
         if T == 1:
             # Fast path: single token (decode).
             eids    = topk_ids[0]                              # (K,)
-            ws      = topk_weights[0].to(hidden_states.dtype)  # (K,)
+            # topk_weights already cast to hidden_states.dtype above (all 3 routing paths)
+            ws      = topk_weights[0]                          # (K,)
 
             # --- EP T=1: skip ghost experts ---
             # In EP mode, ~6/8 experts have weight=0 (non-local).
